Create working example of rat moving at constant speed
This is the first time I've defined any part of my game in data (apart from textures). This architecture is excellent. I tried to do this kind of architecture a long time ago and the problem I was having was I had an object with components and the components needed to be in different managers but still apart of the same entity.

This is much better because Entity doesn't even know that there is an EntityView and EntityController attached to it.

Model-View-Controller is brilliant!