Start fresh
For the past few days I've been experimenting with a few architectures trying to find one that suits my needs. I think I found one.

It was described by Vittorio Romeo in his talk "Implementation of a component-based entity system in modern C++" at CppCon 2015. The code is very similar to that described in the talk. I might deviate as I begin to make the actual game.