Reimplement OrthoMoveDir
This is the original implementation. I didn't like the performance and memory usage of the previous implementation. It wasn't a bottleneck but it was bothering me! This implementation uses 4 bytes of memory whereas the previous used 56 bytes of stack and some heap. This implementation is also less lines of code.