Refactor Camera
After drawing a data dependency graph, (Is that even what its called?) I began to see the flow of the data. I saw a set of functions rather than a set of classes. I'm not sure about CameraToPixels, CameraToMeters and CameraVisible. I intend these classes to model partial application of a function but the user can use the classes unsafely. I'll need to find a way to restrict the interface.