Make CameraZoom::calcPPM purely functional
I wanted to see what functional programming was like and it is rather restrictive. I decided not to allow defining constants within functions meaning I had to create a new function if I wanted to use a value more than once. The compiler for a functional language is allowed to assume that every expression has no side effects so it will memoize everything. C++ compilers don't always do that.

Refactoring the code gave me a very good understanding of what was going on. The code also worked the first time I compiled it because all I did was rearrange the code. I didn't really change it.

This is just an experiment. I'll probably tone down the functionalness later.