Attach SDL_Window to CameraWindowSize rather than use InputSystem
Listening for a window size event with the InputSystem assumes there is only one window. To adjust the system to work with multiple windows, I would have to check the window pointer in the event data. To do that I would have to attach the window to the CameraWindowSize. But if CameraWindowSize has a handle to the window, there's no need to use events because the size could be polled from the handle.