Render game with RenderManager
Next up is ParticleSystem then debug physics then debug camera. I'm not really sure where to put the camera. The camera has its hands in everything. The camera has a debug render function, needs a handle to the window, has the view projection matrix and can track an entity. I might be able to remove the handle to the window and instead pass the window size directly. The debug render can be a job in the RenderManager.

This thing is gonna be huge by the time I finish. I'm learning so much about game development. I kind of wish I started this earlier, rather than spending year 10 (2016) working on a "game engine".