Fix crash after entity despawn
I understand why macros are evil. This crash happened because the ParticleComponent was not removed from the ParticleSystem when the missile entity despawned. This happened because of line 83 of "entity manager.cpp". I defined an empty macro. Macros are evil because it is legal to define an empty macro and compilers don't produce warnings when the parameters of a macro aren't used.

I could have Systems store a std::tuple of the systems. The implementation of EntityManager::destroyComponents would iterate the tuple and remove the component from each of them.