Create droid physics body
There are two types of droids:
 * ground
 * air
and there are three levels for each type:
 * level 0  -  Really dumb. Just moves toward the player
 * level 1  -  A bit smarter. Has ranged weapons
 * level 2  -  Really smart. Has powerful ranged weapons. Good accuracy

The component system is going to be really useful in this situation. I'll have a physics component for each droid type and an AI component for each level. The droids that have ranged weapons will have a spawn component that spawns bullets.

Or maybe an AI component for each combination of type and level because air droids have a second dimension of travel and things are a little more complicated.

I'll figure it out!