Adapt to changes made to Simpleton
This thing is a monster. Holy crap, its huge. But I didn't plan it so it just turned to shit. Pink Pentagon had such potential. So did Buttons but sadly, that didn't turn out as fun as I thought it would be.

I can't decide if trying to improve Simpleton is a waste of time. Every game I make is different so Simpleton has to be generic. But for a game engine to be useful it has to make assumptions and not be generic. So if I need to reuse code I should just copy and paste and maybe consider moving it into Simpleton if the code can be made generic. I don't think Simpleton should be a framework. Simpleton should be a library. Simpleton should be a few things that are useful.

This is a waste of time. I'm going to start my next game right now. Or maybe after lunch.

I could turn these commit messages into a blog that only I would read!